Extract proposal transition service boundary
